Sign In and Sign Up pages edited according to Figma! Also validation added !
